backend/requirements.txt
absl-py==2.1.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.4
django-brotli==0.2.1
django-cors-headers==4.3.1
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-spectacular==0.27.2
executing==2.1.0
flatbuffers==24.12.23
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.1.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.6.0
numba==0.60.0
numpy==2.2.1
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.1
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qrcode==8.0
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.6.0
scipy==1.14.1
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soundfile==0.12.1
soxr==0.5.0.post1
sqlparse==0.5.3
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.6.0
wrapt==1.17.0
